@@ -24,7 +24,7 @@
     "react-dom": "^15.2.1"
   },
   "devDependencies": {
-    "autoprefixer": "^6.5.3",
+    "autoprefixer": "^7.1.1",
     "babel-core": "^6.18.2",
     "babel-eslint": "^6.1.2",
     "babel-loader": "^6.2.4",
@@ -39,13 +39,13 @@
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
     "babel-preset-react": "^6.3.13",
     "chai": "^3.3.0",
-    "css-loader": "^0.25.0",
+    "css-loader": "^0.28.1",
     "del": "^1.1.1",
     "empty": "^0.10.1",
     "eslint": "^3.10.2",
     "eslint-loader": "^1.6.1",
     "eslint-plugin-flowtype": "^2.30.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.1.0",
     "flow-bin": "^0.39.0",
     "gulp": "^3.8.11",
     "gulp-cheerio": "^0.6.2",
@@ -58,18 +58,21 @@
     "gulp-using": "0.0.1",
     "html-webpack-plugin": "^2.24.1",
     "json-loader": "^0.5.4",
-    "node-sass": "^3.11.2",
-    "postcss-loader": "^1.1.1",
+    "node-sass": "^4.5.3",
+    "postcss-loader": "^2.0.5",
+    "postcss-scss": "^1.0.0",
     "react-image-crop": "^2.0.3",
     "resolve-url-loader": "^1.6.0",
     "run-sequence": "^1.0.2",
-    "sass-loader": "^4.0.2",
+    "sass-loader": "^6.0.5",
     "style-loader": "^0.13.1",
-    "webpack": "^1.13.1",
+    "webpack": "^2.5.1",
+    "webpack-dev-server": "^2.4.5",
     "webpack-stream": "^3.2.0"
   },
   "scripts": {
-    "build": "gulp deploy",
+    "dev":  "webpack-dev-server",
+    "deploy": "gulp deploy",
     "flow": "flow; test $? -eq 0 -o $? -eq 2"
   },
   "repository": {
