@@ -1,63 +1,87 @@
-var 	webpack				= require("webpack"),
+const 	webpack				= require("webpack"),
 		path				= require('path'),
 		ExtractTextPlugin	= require('extract-text-webpack-plugin'),
 		HtmlWebpackPlugin	= require('html-webpack-plugin'),
 		autoprefixer		= require('autoprefixer');
 
+const babelPluginsList = [
+	"transform-flow-strip-types",
+	"transform-es2015-arrow-functions",     // allowing arrow functions
+	"check-es2015-constants",               // checking const expressions to be really const
+	"transform-es2015-block-scoping",       // allowing block scope features
+	"transform-es2015-template-literals",   // allow string interpolation
+	"transform-es2015-classes",				// allow class syntax
+	"transform-class-properties",
+	"transform-es2015-parameters",			// transforming default values
+	"transform-es2015-shorthand-properties"
+];
+
+/*
+ * White list of modules which should be transpiled with babel.
+ * This list is required to transpile some modules which are definitely not ES5
+ * @type {Array}
+ */
+const nodeModulesBabelWhiteList = [
+	path.join('node_modules', 'propz')
+];
+
+/** Check whether provided value is in white list */
+nodeModulesBabelWhiteList.isWhiteListed = function(value) {
+	return this.findIndex(whiteItem => value.includes(whiteItem)) !== -1;
+};
+
 module.exports = {
 	entry: "./source/js/init",
 	resolve: {
-		root: [
-			path.resolve('./source')
-		],
-		modulesDirectories: [
+		modules: [
 			'node_modules',
-			'source/js',
-			'source'
-		]
+			path.resolve(__dirname, 'source/js'),	// do we need both?
+			path.resolve(__dirname, 'source')
+		],
+		alias: {
+			director: path.resolve(__dirname, 'node_modules/director/build/director')
+		}
 	},
 	stats: {
 		children: false // not showing chatty logs from Child plugin
 	},
 	module: {
-		preLoaders: [
+		rules: [
 			{
 				test: /\.js$/,
-				exclude: /(node_modules|bower_components)/,
-				loader: 'eslint'
+				exclude: value => {
+					/* exclude if not whitelisted and in node_modules */
+					if(nodeModulesBabelWhiteList.isWhiteListed(value)) {
+						return false;
+					} else {
+						return /(node_modules)/.test(value);
+					}
+				},
+				use: [
+					{ loader: 'eslint-loader' },
+					{ loader: 'babel-loader', options: {
+							"presets": ["react"],
+							"plugins": babelPluginsList
+					}}
+				]
+			},
+			{
+				test: /\.scss$/,
+				loader: ExtractTextPlugin.extract({
+					fallback: 'style-loader',
+					use: ['css-loader', 'postcss-loader','sass-loader']
+				})
 			}
 		],
-		loaders: [
-			{
-				test: /\.js$/,
-				exclude: /(node_modules|bower_components)/,
-				loader: 'babel',
-				query: {
-					"presets": ["react"],
-					"plugins": [
-						"transform-flow-strip-types",
-						"transform-es2015-arrow-functions",     // allowing arrow functions
-						"check-es2015-constants",               // checking const expressions to be really const
-						"transform-es2015-block-scoping",       // allowing block scope features
-						"transform-es2015-template-literals",   // allow string interpolation
-						"transform-es2015-classes",				// allow class syntax
-						"transform-class-properties",
-						"transform-es2015-parameters",			// transforming default values
-						"transform-es2015-shorthand-properties"
-					]
-				}
-			}, {
-				test: /\.scss$/,
-				loader: ExtractTextPlugin.extract('css!postcss!sass')
-			}, {
-				include: /\.json$/,
-				loaders: ["json-loader"]
-			}
-		]
 	},
-	postcss: [ autoprefixer({ browsers: ['last 2 versions'] }) ],
+	devtool: 'source-map',
 	plugins: [
-		new ExtractTextPlugin('styles.css', {
+		new webpack.optimize.UglifyJsPlugin({
+			mangle:		false,	// I'm not sure if mangling can be enabled safely. So disabling it for a while
+			sourceMap:	true
+		}),
+		new ExtractTextPlugin({
+			filename: 'styles.css',
 			allChunks: true
 		}),
 		new HtmlWebpackPlugin({
@@ -71,5 +95,8 @@ module.exports = {
 		publicPath: 'dist/',					// specifies the public URL address of the output files when referenced in a browser
 		path: 		path.resolve('./dist'),		// storing all results in this folder
 		filename: 	'bundle.js'					// with names like this
+	},
+	devServer: {
+		disableHostCheck: true
 	}
 };
